// src/models/User.ts
interface IUser {
  name?: string;
  email?: string;
  password?: string;
  passwordHash?: string;
  companyName?: string;
  businessType?: string;
  rememberMe?: string;
  planID?: number;
  locationsCount?: string;
  townCity?: string;
  stateCountyRegion?: string;
  googleEmail?: string;
  googlePassword?: string;
  shareAddress?: string;
  shareFirstName?: string;
  shareLastName?: string;
  gmbGroup?: string;
  gmbLocation?: string;
  locationName?: string;
  locationAddress?: string;
  subscriptionId?: number;
  userEmail?: string;
  cid?: number;
  uid?: number;
  userType?: string;
}

export { IUser };